updated uninstallation script

+++ v5/bin/gmetrics_agent_uninstall.sh
@@ -53,6 +53,9 @@ while read line; do echo "[`date +"%Y-%m-%dT%H:%M:%S,%N" | rev | cut -c 7- | rev
 
 disable_service () {
 
+if [ -f /lib/systemd/system/gmetrics-agent.service ];then
+
+
 echo "#################################################" | log
 echo "Disabling gmetrics-agent service" | log
 sudo systemctl disable gmetrics-agent | log
@@ -65,6 +68,13 @@ echo "gmetrics-agent service successfully removed." | log
 echo "#################################################" | log
 echo "Daemon reloading to make changes" | log
 sudo systemctl daemon-reload | log
+
+else
+
+echo "gmetrics-agent.service is not present. Could not remove. Exiting..." | log 
+exit 1;
+
+fi
 }
 
 # To remove /etc/logrotate.d/gmetrics-agent logrotate file
@@ -74,9 +84,17 @@ remove_logrotate () {
 
 echo "#################################################" | log
 echo "Removing logrotate file" | log
+
+if [ -f /etc/logrotate.d/gmetrics-agent ]; then
+echo "#################################################" | log
 sudo rm -rvf /etc/logrotate.d/gmetrics-agent | log
 echo "#################################################" | log
 echo "Removed "/etc/logrotate.d/gmetrics-agent" file" | log
+
+else
+echo "#################################################" | log
+echo "No /etc/logrotate.d/gmetrics-agent file found. Could not remove..!" | log
+fi
 }
 
 # To remove port from /etc/services
@@ -102,16 +120,18 @@ echo "#################################################" | log
 echo "Emptying "/groots/tmp/" directory" | log
 rm -rvf /groots/tmp/* | log
 rm -rvf /groots/tmp/.svn/ | log
+
+if [ -d /groots/metrics  ]; then
 echo "#################################################" | log
 echo "Removing "/groots/metrics" directory" | log
 rm -rvf /groots/metrics | log
-echo "Verifing deletion of "/groots/metrics" directory.." | log
+echo "#################################################" | log
+echo ""/groots/metrics" directory is present and removed successfully" | log
 
-if [ ! -d /groots/metrics ]; then
-echo "/groots/metrics is completely removed." | log
 else
-echo "/groots/metrics could not be removed. " | log
-exit 1;
+echo "#################################################" | log
+echo "No "/groots/metrics" directory found. Could not remove..!!" | log
+
 fi
 
 }
@@ -123,16 +143,17 @@ remove_user () {
 
 echo "#################################################" | log
 echo "Removing groots user" | log
-userdel -r groots
-if [ ! -d /home/groots/ ]; then
+getent passwd groots   > /dev/null
+
+if [ `echo $?` -eq 0 ]; then
+userdel -r groots > /dev/null
 echo "#################################################" | log
-echo "groots user home directory is not present" | log
+echo "Removed groots user and home directory" | log
 else
 echo "#################################################" | log
-rm -rf /home/groots | log
-echo "Removed groots home directory" | log
+echo "groots user not found..!!" | log
+exit 1;
 fi
-
 }
 
 # To remove firewall port for gmetrics-agent
@@ -166,9 +187,17 @@ remove_sudoers_entry () {
 
 echo "#################################################" | log
 echo "Removing sudoers entry for gmetrics-agent" | log
-rm -rvf /etc/sudoers.d/gmetrics-agent | log
-echo "removed "/etc/sudoers.d/gmetrics-agent" " | log
 
+if [ -f /etc/sudoers.d/gmetrics-agent ]; then
+echo "#################################################" | log
+rm -rvf /etc/sudoers.d/gmetrics-agent | log
+echo "Removed "/etc/sudoers.d/gmetrics-agent" " | log
+
+else
+echo "#################################################" | log
+echo "No entry for gmetrics-agent in sudoers found.." | log
+
+fi
 }
 
 # Function calling
@@ -176,22 +205,22 @@ echo "removed "/etc/sudoers.d/gmetrics-agent" " | log
 
 while true
 do
-	echo "#######################################################" | log
-	echo "Gmetrics agent uninstallation starting at [`date`]." | log
+        echo "#######################################################" | log
+        echo "Gmetrics agent uninstallation starting at [`date`]." | log
 
-	disable_service
-	remove_firewall_entry
-	remove_logrotate
-	remove_service_port
-	remove_user
-	remove_sudoers_entry
-	remove_groots_directory
+        disable_service
+        remove_firewall_entry
+        remove_logrotate
+        remove_service_port
+        remove_user
+        remove_sudoers_entry
+        remove_groots_directory
 
-	echo "#######################################################" | log
-	echo "Gmetrics Agent is successfully uninstalled." | log
-	echo "Gmetrics Agent uninstallation is completed at [`date`]." | log
-	echo "#######################################################" | log
-		
+        echo "#######################################################" | log
+        echo "Gmetrics Agent is successfully uninstalled." | log
+        echo "Gmetrics Agent uninstallation is completed at [`date`]." | log
+        echo "#######################################################" | log
+
 break
 done
